ocr-api-develop/requirements.txt
alembic==1.13.1
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
dnspython==2.6.1
docstring-parser==0.16
email-validator==2.1.1
Faker==24.7.1
fastapi==0.110.1
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.1.0
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pluggy==1.4.0
psycopg2==2.9.9
py==1.11.0
pycparser==2.22
pydantic==2.6.4
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
zipp==3.18.1
pydantic-settings==2.2.1
httpx==0.27.0
minio==7.2.5
ndg-httpsclient==0.5.1
pyopenssl==24.1.0
pyasn1==0.6.0
pip-system-certs==4.0
pillow==10.3.0
aiohttp==3.9.5
asyncpg==0.29.0
redis[hiredis]==5.0.5
aiokafka==0.10.0
confluent_kafka==2.4.0
elastic-apm==6.22.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.8.0
transformers==4.41.0
torch~=2.3.0
pydantic-settings~=2.1.0
python-dotenv~=1.0.0
pypdfium2~=4.25.0
opencv-python~=4.9.0.80
tabulate~=0.9.0
filetype~=1.2.0
ftfy~=6.1.3
pymupdf~=1.23.8
snakeviz~=2.2.0
datasets~=2.16.1
rapidfuzz~=3.6.1
arabic-reshaper~=3.0.0
playwright~=1.41.2
paddleocr~=2.0.1
paddlepaddle~=2.6.1